+++ app/controllers/documentation/pages_controller.rb
@@ -6,7 +6,7 @@ module Documentation
     before_action :set_workspace
     before_action :set_root_pages, except: %i[index]
     before_action :set_page, only: %i[show edit update destroy]
-    before_action :set_documents, only: %i[show edit]
+    before_action :set_files, only: %i[show edit]
 
     def index
       @pages = @workspace.pages
@@ -82,8 +82,8 @@ module Documentation
       @root_pages = @workspace.pages.roots.include_tree
     end
 
-    def set_documents
-      @documents = @page.documents
+    def set_files
+      @files = @page.files
     end
   end
 end
