+++ app/views/documentation/files/index.html.erb
@@ -1,16 +1,5 @@
-<%= render(Bali::Tabs::Component.new) do |c| %>
-  <% c.tab(title: t('documentation.page.images'), icon: 'image', active: true) do %>
-    <%= turbo_frame_tag 'page_images' do %>
-      <%= render 'documentation/images/form', page: @page %>
-      <hr />
-      <%= render 'documentation/images/grid', images: @images, page: @page %>
-    <% end %>
-  <% end %>
-  <% c.tab(title: t('documentation.page.documents'), icon: 'file-signature') do %>
-    <%= turbo_frame_tag 'page_documents' do %>
-      <%= render 'documentation/documents/form', page: @page %>
-      <hr />
-      <%= render 'documentation/documents/list', documents: @documents, page: @page %>
-    <% end %>
-  <% end %>
+<%= turbo_frame_tag 'page_files' do %>
+  <%= render 'form', page: @page %>
+  <hr />
+  <%= render 'documentation/pages/files', files: @files, page: @page %>
 <% end %>
