--- app/views/documentation/pages/_documents.html.erb
@@ -1,1 +1,0 @@
-<%= render 'documentation/documents/list', documents: @documents, page: @page %>
