+++ app/views/documentation/pages/_files.html.erb
@@ -1,0 +1,1 @@
+<%= render 'documentation/files/list', files: @files, page: @page %>
